+++ sphinx/source/changelog.rst
@@ -12,6 +12,8 @@ Changelog
 * Added the ``--sim-load-config`` option to :ref:`tap2sna.py` (to set the value
   of a ``--sim-load`` configuration option: ``accelerator``, ``fast-load``,
   ``pause``, ``polarity``)
+* Added the ``--tape-start`` option to :ref:`tap2sna.py` (to start the tape at
+  a specific block number)
 * Added the ``--trace`` option to :ref:`tap2sna.py` (to log the instructions
   executed during a simulated LOAD to a file)
 * Added the ``--interrupts`` option to :ref:`trace.py` (to enable the execution